Dev (#13) (#17)
* added cli submodule

* use dev branch of execDAT-CLI

* use main branch for checkout of cli in main

* add index and valueproposition

* Template for the ADRs

We adapted an ADR template we found online. Some changes are unique to our template: We are using the date as prefix in the ADR folder for natural ordering. Also, we require a link to the corresponding ADR in deprecated and superseded ADRs.

* add operator as submodule

* add architecture and local k3d setup

* Template for the ADRs (#6)

We adapted an ADR template we found online. Some changes are unique to our template: We are using the date as prefix in the ADR folder for natural ordering. Also, we require a link to the corresponding ADR in deprecated and superseded ADRs.

* public vs private ADR



* testing for gpg signing

* accepted public-vs-private-data ADR

* proposed branch-naming ADR

* fix: use the port of the devel (#7)

* Start CICD ADR

* test signed commit

* Fix commit sign

* Finish CICD ADR

* Some small changes to cicd ADR

* Start lecture 9 readme



* finish lecture 9



---------

Co-authored-by: Naexon <a.woda@outlook.com>
Co-authored-by: Naexon <75274749+Naexon@users.noreply.github.com>
Co-authored-by: Thomas Weber <kontakt@weber-thomas.at>
Co-authored-by: Daniel Hofstätter <daniel.hofstaetter@tuwien.ac.at>
Co-authored-by: Thomas Weber <websi96@users.noreply.github.com>
Co-authored-by: Naexon <Naexon@users.noreply.github.com>